Fixed stuff
Accidentally lost the Example scene in the restructure, its fine, i had to redo the UI for mobile portrait mode anyway, and that was better done from scratch than by fighting with the old stuff. also made it easier to use prefabs from the UI pack - note to self. next time you make a game, import the UI pack first,  and use it from the begining. this is tedious.